// src/types/Config.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Config = { ttpmacro_path: string, server_data_api: string, user_data_api: string, default_login_user: string, default_login_su: string, default_login_win: string, default_login_oc: string, };